+++ resources/template/helpers.h
@@ -1,0 +1,6 @@
+#ifndef HELPERS_H
+#define HELPERS_H
+
+bool strings_equal(const char *str1, const char *str2);
+
+#endif