+++ resources/template/print.h
@@ -1,0 +1,10 @@
+#ifndef PRINT_H
+#define PRINT_H
+
+#include "types.h"
+
+void print_string(const char* x);
+void print_record(value_record x);
+void print_value(value x);
+
+#endif