+++ resources/template/reduce.h
@@ -1,0 +1,9 @@
+#ifndef REDUCE_H
+#define REDUCE_H
+
+#include <stdbool.h>
+#include "types.h"
+
+bool reduce(value* x);
+
+#endif